Implementing  (partially) resource substitution
Partially working in that it is able to sub something and work up the
crafting tree. However, it cannot backtrack all the way down. Will need
to implement something for keeping track of "tried" hypotheses -- the
ones that seem to work, but later the agent finds out that they may not.
Maybe MCTS type thing with hypotheses, and backtracking.